# task-manager-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TaskManager API",
    "version": "1.0.0",
    "description": "A simple API for managing users and tasks."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/User" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "summary": "Update a user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/User" } }
          }
        },
        "responses": {
          "200": { "description": "User updated" },
          "400": { "description": "Validation error" }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User deleted" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Task" } }
          }
        },
        "responses": {
          "201": { "description": "Task created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/tasks/{id}": {
      "put": {
        "summary": "Update a task",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } }
        },
        "responses": {
          "200": { "description": "Task updated" },
          "400": { "description": "Validation error" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Task deleted" },
          "500": { "description": "Server error" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": ["name", "email", "age"],
        "properties": {
          "name": { "type": "string", "example": "Alice Johnson" },
          "email": { "type": "string", "example": "alice@example.com" },
          "age": { "type": "integer", "example": 28 }
        }
      },
      "Task": {
        "type": "object",
        "required": ["title", "description", "assignedTo"],
        "properties": {
          "title": { "type": "string", "example": "Complete project" },
          "description": { "type": "string", "example": "Finish CRUD operations" },
          "assignedTo": { "type": "string", "example": "607d1b2f8c1b2c0015f4b8c1" },
          "status": { "type": "string", "enum": ["Pending", "In Progress", "Completed"], "example": "Pending" }
        }
      }
    }
  }
}